+++ zone/lua_group.h
@@ -34,7 +34,7 @@ public:
 	bool IsGroupMember(const char* name);
 	bool IsGroupMember(Lua_Mob c);
 	void CastGroupSpell(Lua_Mob caster, int spell_id);
-	void SplitExp(uint32 exp, Lua_Mob other);
+	void SplitExp(uint64 exp, Lua_Mob other);
 	void GroupMessage(Lua_Mob sender, const char* message);
 	void GroupMessage(Lua_Mob sender, int language, const char* message);
 	uint32 GetTotalGroupDamage(Lua_Mob other);
